Create farcaster project from farcaster app
