+++ apps/connect/components/projects/edit/EditProjectPage.tsx
@@ -2,13 +2,16 @@
 
 import { log } from '@charmverse/core/log';
 import { PageWrapper } from '@connect-shared/components/common/PageWrapper';
+import { ProjectForm } from '@connect-shared/components/project/ProjectForm';
 import type { LoggedInUser } from '@connect-shared/lib/profile/getCurrentUserAction';
 import type { ConnectProjectDetails } from '@connect-shared/lib/projects/fetchProject';
 import type { FormValues, ProjectCategory } from '@connect-shared/lib/projects/form';
 import { schema } from '@connect-shared/lib/projects/form';
 import { yupResolver } from '@hookform/resolvers/yup';
-import { Typography } from '@mui/material';
+import { Button, Stack, Typography } from '@mui/material';
 import { Box } from '@mui/system';
+import { ConnectApiClient } from 'apiClient/apiClient';
+import Link from 'next/link';
 import { useRouter } from 'next/navigation';
 import { useAction } from 'next-safe-action/hooks';
 import { useState } from 'react';
@@ -20,7 +23,8 @@ import { actionEditProject } from 'lib/projects/editProjectAction';
 import { AddProjectMembersForm } from '../components/AddProjectMembersForm';
 import type { ProjectDetailsProps } from '../components/ProjectDetails';
 import { ProjectDetails } from '../components/ProjectDetails';
-import { ProjectForm } from '../components/ProjectForm';
+
+const connectApiClient = new ConnectApiClient();
 
 export function EditProjectPage({ user, project }: { user: LoggedInUser; project: ConnectProjectDetails }) {
   const [showTeamMemberForm, setShowTeamMemberForm] = useState(false);
@@ -68,13 +72,31 @@ export function EditProjectPage({ user, project }: { user: LoggedInUser; project
   if (!showTeamMemberForm) {
     return (
       <PageWrapper>
-        <ProjectForm
-          control={control}
-          isValid={isValid}
-          onNext={() => {
-            setShowTeamMemberForm(true);
-          }}
-        />
+        <ProjectForm uploadImageFn={connectApiClient.uploadImage} control={control} />
+        <Stack
+          justifyContent='space-between'
+          flexDirection='row'
+          position='sticky'
+          bottom='0'
+          bgcolor='background.default'
+          py={2}
+        >
+          <Link href='/profile' passHref>
+            <Button size='large' color='secondary' variant='outlined'>
+              Cancel
+            </Button>
+          </Link>
+          <Button
+            data-test='project-form-confirm-values'
+            size='large'
+            disabled={!isValid}
+            onClick={() => {
+              setShowTeamMemberForm(true);
+            }}
+          >
+            Next
+          </Button>
+        </Stack>
       </PageWrapper>
     );
   }
