+++ apps/connect/tsconfig.json
@@ -117,7 +117,10 @@
     "**/*.tsx",
     ".next/types/**/*.ts",
     "../../@connect-shared/components/common/NotFoundPage.tsx",
-    "../../@connect-shared/components/common/FarcasterCard.tsx"
+    "../../@connect-shared/components/common/FarcasterCard.tsx",
+    "../../@connect-shared/components/project/ProjectForm.tsx",
+    "../../@connect-shared/hooks/useS3UploadInput.ts",
+    "../../@connect-shared/hooks/useFilePicker.ts"
   ],
   "exclude": [
     "../../node_modules",
